chore: better clean script and readme update

+++ scripts/clean.sh
@@ -1,12 +1,25 @@
 #!/bin/sh
+ROOT_BUILD_DIR=ruby-build
 
-rm -f CMakeCache.txt Makefile cmake_install.cmake
-rm -rf CMakeFiles
-rm -rf ruby-build/CMakeFiles
-for item in "ruby-build"/*
-do
-  [ -d "$item" ] && rm -rf $item/build_dir/*
-done
-rm -f ruby-build/cmake_install.cmake ruby-build/Makefile
+rm -rf $ROOT_BUILD_DIR/CMakeFiles
 
-rm -rf target/
+TARGETS="$@"
+
+clear_folder() {
+  local folder="$1"
+  [ -d "$folder" ] && rm -rf "$folder/build_dir/"* && rm -rf "$folder/build_dir/".*
+}
+
+if [ -z "$TARGETS" ]; then
+  rm -f CMakeCache.txt Makefile cmake_install.cmake
+  rm -rf CMakeFiles
+  for item in "$ROOT_BUILD_DIR"/*; do
+    clear_folder "$item"
+  done
+  rm -f $ROOT_BUILD_DIR/cmake_install.cmake $ROOT_BUILD_DIR/Makefile
+  rm -rf target/
+else
+  for target in $TARGETS; do
+    clear_folder "$ROOT_BUILD_DIR/$target"
+  done
+fi;
